Better Set apexType Matching
I noticed there was highlighting on properties when {get; set;}
was used. This was pretty ugly since the get keyword was not
highlighted.

The change is removing Set from the keword match and instead
using regex to find Set followed by a < char to indicate that
it’s a the collection that we’re looking for. I then offset
the match end by 3 chars (length of the word “set”) from the
start of the match.

The only caveats here are that if you use a class method of
Set, the word would not be highlighted. But that’s moot since
I don’t believe there are any Set Class Methods.