Add an new match syntax for 'new'
Previous keyword syntax was highlighting every time `new` showed up in
the file. This was annoying for things like `Trigger.new` where `new` is
not functioning as an operator, but rather a normal property.

This patch removes `new` from the `syn keyword` and instead uses `syn
match` to select new only if it does not follow a `.` as in `.new`.