Fix word ending on new matching
While I resolved incorrectly highting `Trigger.new`, I ended up creating
a problem with variables that start with the word new because I forgot
to specify a word end as part of the match. The result was that
`newUser` would have the `new` be highlighted.

This patch resolves the issue.